pyproject.toml
[tool.poetry]
name = "mlops_project"
version = "0.1.0"
description = "Template MLOps End-to-End avec training/predict en containers sur ECS"
authors = ["Ton Nom <ton.email@example.com>"]
readme = "README.md"
packages = [{ include = "mlops_project", from = "src" }]

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.3"
boto3 = "^1.37.38"
requests = "^2.31.0"
scikit-learn = "^1.4.0"
mlflow = "^2.11.1"
python-dotenv = "^1.0.0"
poetry-core = "^2.1.2"
seaborn = "^0.13.2"
mkdocs = "^1.6.1"
mkdocs-material = "^9.6.12"
mkdocstrings = {extras = ["python"], version = "^0.29.1"}
pymysql = "^1.1.1"

[tool.poetry.group.dev.dependencies]
jupyterlab = "^4.1.0"
ipykernel = "^6.29.0"
notebook = "^7.1.0"
black = "^24.3.0"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
